abu_engine/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
skyfield>=1.39.0
numpy>=1.21.0
pandas>=1.3.0
python-multipart>=0.0.5
pyswisseph>=2.10.3.2